세번째 커밋

+++ src/main/java/kr/ac/hansung/cse/repo/ProductRepository.java
@@ -6,6 +6,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import kr.ac.hansung.cse.model.Product;
 
-public interface ProductRepository extends CrudRepository<Product, Long>{
+//id가 int 타입이라 Integer로 주어야 한다
+public interface ProductRepository extends CrudRepository<Product, Integer>{
 	List<Product> findByCategory(String category);
 }
